% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{add_axis_line}
\alias{add_axis_line}
\title{Align axis line to zero and mark darker}
\usage{
add_axis_line(axis = c("x", "y"))
}
\arguments{
\item{axis}{Apply to "x" or "y" axis}
}
\description{
Align axis line to zero and mark darker
}
