% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{add_zero_line}
\alias{add_zero_line}
\title{Add darker line to zero intercept}
\usage{
add_zero_line(axis = c("y", "x"))
}
\arguments{
\item{axis}{Apply to "x" or "y" axis.  Defaults to "y".}
}
\description{
Add darker line to zero intercept
}
