% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{flip_gridlines}
\alias{flip_gridlines}
\title{Flip default major gridlines from horizontal to vertical}
\usage{
flip_gridlines(
  color = "#cbcdcc",
  linewidth = 0.05,
  size = lifecycle::deprecated()
)
}
\arguments{
\item{color}{Gridline color}

\item{linewidth}{Gridline width}

\item{size}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{linewidth} instead.}
}
\description{
Flip default major gridlines from horizontal to vertical
}
