% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{get_logo}
\alias{get_logo}
\title{Get a logo for use in a ggplot2 plot}
\usage{
get_logo(type = c("hda", "hfv"), width = 100)
}
\arguments{
\item{type}{one of "hda" or "hfv"}

\item{width}{Image width in pixels; defaults to 100}
}
\value{
a path to an image to be used in ggplot2 plots
}
\description{
Get a logo for use in a ggplot2 plot
}
