% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{hda_pal_discrete}
\alias{hda_pal_discrete}
\title{Generate a discrete HDA color palette}
\usage{
hda_pal_discrete(direction = 1)
}
\arguments{
\item{direction}{If -1, reverse the palette; defaults to 1}
}
\value{
n colors (generally passed to ggplot2)
}
\description{
Generate a discrete HDA color palette
}
