% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdatools-package.R
\docType{package}
\name{hdatools-package}
\alias{hdatools}
\alias{hdatools-package}
\title{hdatools: Tools and Themes for 'HDAdvisors' Projects}
\description{
Tools and workflows for HDAdvisors and HousingForward Virginia Projects. Functions include general analytical utilities and custom ggplot2 themes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hdadvisors/hdatools}
  \item Report bugs at \url{https://github.com/hdadvisors/hdatools/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Knopf \email{jonathan@hdadvisors.net}

Authors:
\itemize{
  \item Kyle Walker \email{kyle@walker-data.com}
}

}
\keyword{internal}
