% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{hfv_pal_discrete}
\alias{hfv_pal_discrete}
\title{Generate a discrete HFV color palette}
\usage{
hfv_pal_discrete(direction = 1, repeat_pal = FALSE)
}
\arguments{
\item{direction}{If -1, reverse the palette; defaults to 1}

\item{repeat_pal}{If TRUE, repeat the palette enough times to account for all discrete values}
}
\value{
n colors (generally passed to ggplot2)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{scale_color_hfv()}/\code{scale_fill_hfv()}
directly instead of calling this palette generator.
}
