% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{markdown_wrap_gen}
\alias{markdown_wrap_gen}
\title{Generate a function to wrap and format facet labels with markdown}
\usage{
markdown_wrap_gen(width = 25)
}
\arguments{
\item{width}{An integer specifying the maximum number of characters
before wrapping the text. Default is 25.}
}
\value{
A function that takes a vector of labels and returns a list of
  wrapped and formatted labels.
}
\description{
This function creates a labeller function for use with ggplot2 facets.
It wraps long labels to a specified width and formats them as markdown,
which allows them to be rendered correctly when using ggtext::element_markdown()
in themes.
}
\examples{
\dontrun{
library(ggplot2)
library(hdatools)

ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  facet_wrap(~vs, labeller = markdown_wrap_gen(width = 20)) +
  theme_hda()
}

}
