% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{multiply}
\alias{multiply}
\title{Multiply two numbers}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{the first number}

\item{y}{the second number}
}
\value{
the product
}
\description{
Multiply two numbers
}
