% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{publish_plot}
\alias{publish_plot}
\title{Create dynamic graphic from plot object when document rendered as HTML}
\usage{
publish_plot(plot)
}
\arguments{
\item{plot}{ggplot object}
}
\description{
Create dynamic graphic from plot object when document rendered as HTML
}
