% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_gradient_hda}
\alias{scale_color_gradient_hda}
\alias{scale_colour_gradient_hda}
\title{HDA-branded 4-color continuous color scale}
\usage{
scale_color_gradient_hda(
  ...,
  colors = .brands$hda$gradient,
  values = NULL,
  space = "Lab",
  na.value = .brands$hda$na_color,
  guide = "colorbar"
)

scale_colour_gradient_hda(
  ...,
  colors = .brands$hda$gradient,
  values = NULL,
  space = "Lab",
  na.value = .brands$hda$na_color,
  guide = "colorbar"
)
}
\arguments{
\item{...}{Other arguments passed on to continuous_scale()}

\item{colors}{Vector of colors}

\item{values}{If colors should not be evenly positioned along the gradient, this vector gives the position (between 0 and 1) for each color in the vector}

\item{space}{Color space in which to calculate gradient. Must be "Lab" - other values deprecated}

\item{na.value}{Default color for NA values (#cfcfd0, HDA Light Gray)}

\item{guide}{Legend representation for scale}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
