% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_hfv}
\alias{scale_color_hfv}
\alias{scale_colour_hfv}
\title{HFV-branded discrete color scale}
\usage{
scale_color_hfv(direction = 1, repeat_pal = FALSE, ...)

scale_colour_hfv(direction = 1, repeat_pal = FALSE, ...)
}
\arguments{
\item{direction}{If -1, reverse the scale (defaults to 1)}

\item{repeat_pal}{If TRUE, repeat the palette enough times to account for all discrete values}

\item{...}{Additional arguments passed to ggplot2::discrete_scale()}
}
\value{
A \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} object.
}
\description{
HFV-branded discrete color scale
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(wt, mpg, color = factor(cyl))) +
  geom_point() +
  scale_color_hfv()

}
