% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_gradient_pha}
\alias{scale_fill_gradient_pha}
\title{PHA-branded 4-color continuous fill scale}
\usage{
scale_fill_gradient_pha(
  ...,
  colors = .brands$pha$gradient,
  values = NULL,
  space = "Lab",
  na.value = .brands$pha$na_color,
  guide = "colorbar"
)
}
\arguments{
\item{...}{Other arguments passed on to continuous_scale()}

\item{colors}{Vector of colors}

\item{values}{If colors should not be evenly positioned along the gradient, this vector gives the position (between 0 and 1) for each color in the vector}

\item{space}{Color space in which to calculate gradient. Must be "Lab" - other values deprecated}

\item{na.value}{Default color for NA values (#e2e4e3, PHA Light Gray)}

\item{guide}{Legend representation for scale}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
