% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_hfv}
\alias{scale_fill_hfv}
\title{HFV-branded discrete fill scale}
\usage{
scale_fill_hfv(direction = 1, repeat_pal = FALSE, ...)
}
\arguments{
\item{direction}{If -1, reverse the scale (defaults to 1)}

\item{repeat_pal}{If TRUE, repeat the palette enough times to account for all discrete values}

\item{...}{Additional arguments passed to ggplot2::discrete_scale()}
}
\value{
A \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} object.
}
\description{
HFV-branded discrete fill scale
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_hfv()

}
