% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{flip_gridlines}
\alias{flip_gridlines}
\title{Flip default major gridlines from horizontal to vertical}
\usage{
flip_gridlines(color = "#e2e4e3", size = 0.05)
}
\arguments{
\item{color}{Gridline color}

\item{size}{Gridline size}
}
\description{
Flip default major gridlines from horizontal to vertical
}
