% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_helpers.R
\name{get_logo}
\alias{get_logo}
\title{Get a logo for use in a ggplot2 plot}
\usage{
get_logo(type = c("hda", "hfv"), width = 100)
}
\arguments{
\item{type}{one of "hda" or "hfv"}

\item{width}{Image width in pixels; defaults to 100}
}
\value{
A length-one character (glue) string containing an HTML \verb{<img>} tag
that points at the installed logo file, sized to \code{width}. Intended for use
in ggtext-rendered plot elements such as a markdown title or caption.
}
\description{
Get a logo for use in a ggplot2 plot
}
