% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_gradient_hda}
\alias{scale_color_gradient_hda}
\title{HDA-branded 4-color continuous color scale}
\usage{
scale_color_gradient_hda(
  ...,
  colors = c("#445ca9", "#8baeaa", "#e9ab3f", "#e76f52"),
  values = NULL,
  space = "Lab",
  na.value = "cfcfd0",
  guide = "colorbar"
)
}
\arguments{
\item{...}{Other arguments passed on to continuous_scale()}

\item{colors}{Vector of colors}

\item{values}{If colors should not be evenly positioned along the gradient, this vector gives the position (between 0 and 1) for each color in the vector}

\item{space}{Color space in which to calculate gradient. Must be "Lab" - other values depreciated}

\item{na.value}{Default color for NA values (#cfcfd0, HDA Light Gray)}

\item{guide}{Legend representation for scale}
}
\description{
HDA-branded 4-color continuous color scale
}
