% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_hda}
\alias{scale_color_hda}
\title{HDA-branded discrete color scale}
\usage{
scale_color_hda(direction = 1, ...)
}
\arguments{
\item{direction}{If -1, reverse the scale (defaults to 1)}
}
\description{
HDA-branded discrete color scale
}
