% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_hfv}
\alias{scale_color_hfv}
\title{HFV-branded discrete color scale}
\usage{
scale_color_hfv(direction = 1, ...)
}
\arguments{
\item{direction}{If -1, reverse the scale (defaults to 1)}
}
\description{
HFV-branded discrete color scale
}
