% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_pha}
\alias{scale_color_pha}
\title{PHA-branded discrete color scale}
\usage{
scale_color_pha(direction = 1, repeat_pal = FALSE, ...)
}
\arguments{
\item{direction}{If -1, reverse the scale (defaults to 1)}

\item{repeat_pal}{If TRUE, repeat the palette enough times to account for all discrete values}
}
\description{
PHA-branded discrete color scale
}
