% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_vha}
\alias{scale_color_vha}
\alias{scale_colour_vha}
\title{VHA-branded discrete color scale}
\usage{
scale_color_vha(direction = 1, repeat_pal = FALSE, ...)

scale_colour_vha(direction = 1, repeat_pal = FALSE, ...)
}
\arguments{
\item{direction}{If -1, reverse the scale (defaults to 1)}

\item{repeat_pal}{If TRUE, repeat the palette enough times to account for all discrete values}

\item{...}{Additional arguments passed to ggplot2::discrete_scale()}
}
\description{
VHA-branded discrete color scale
}
