% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_hfv}
\alias{scale_fill_hfv}
\title{HFV-branded discrete fill scale}
\usage{
scale_fill_hfv(direction = 1, repeat_pal = FALSE, ...)
}
\arguments{
\item{direction}{If -1, reverse the scale (defaults to 1)}

\item{repeat_pal}{If TRUE, repeat the palette enough times to account for all discrete values}
}
\description{
HFV-branded discrete fill scale
}
