% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_pha}
\alias{scale_fill_pha}
\title{PHA-branded discrete fill scale}
\usage{
scale_fill_pha(direction = 1, ...)
}
\arguments{
\item{direction}{If -1, reverse the scale (defaults to 1)}
}
\description{
PHA-branded discrete fill scale
}
