% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_hfv}
\alias{theme_hfv}
\title{Use a HousingForward Virginia-branded ggplot2 theme}
\usage{
theme_hfv(base_size = 10, base_family = "Open Sans", flip_gridlines = FALSE)
}
\arguments{
\item{base_size}{The base size of text elements; defaults to 14}

\item{base_family}{The base font family; defaults to "Open Sans"}

\item{flip_gridlines}{Orientation of major gridlines; defaults to FALSE for y-axis}
}
\description{
Use a HousingForward Virginia-branded ggplot2 theme
}
