% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_hfv}
\alias{theme_hfv}
\title{Use a HousingForward Virginia-branded ggplot2 theme}
\usage{
theme_hfv(base_size = 12, base_family = "Open Sans")
}
\arguments{
\item{base_size}{The base size of text elements; defaults to 14}

\item{base_family}{The base font family; defaults to "Open Sans"}
}
\description{
Use a HousingForward Virginia-branded ggplot2 theme
}
