% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_pha}
\alias{theme_pha}
\title{Use a PHA-branded ggplot2 theme}
\usage{
theme_pha(base_size = 10, base_family = "Noto Sans")
}
\arguments{
\item{base_size}{The base size of text elements; defaults to 14}

\item{base_family}{The base font family; defaults to "Noto Sans"}
}
\description{
Use a PHA-branded ggplot2 theme
}
